HtmlUi: TableHeaderMenu should be a popup
Advantages: 
- May be displayed if table is small
- May be displayed if menu is opened at the right side of the screen
- Gets opened on top if there is not space on the bottom
- Repositions when table gets horizontally scrolled

+++ org.eclipse.scout.rt.ui.html/src/main/js/jquery/jquery-scout.js
@@ -173,11 +173,11 @@
 
   $.pxToNumber = function(pixel) {
     if (!pixel) {
-      // parseInt would return NaN if pixel is '' or undefined
+      // parseFloat would return NaN if pixel is '' or undefined
       return 0;
     }
-    // parseInt ignores 'px' and just extracts the number
-    return parseInt(pixel, 10);
+    // parseFloat ignores 'px' and just extracts the number
+    return parseFloat(pixel, 10);
   };
 
   // attr and class handling for svg
